Fix: Move tests into appropriate namespace
